requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
redis==5.2.1
pydantic-settings==2.6.1
httpx==0.27.2
aiomysql==0.2.0
PyJWT[crypto]==2.10.1

# 해외출장 자동화 모듈
playwright==1.44.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
sqlalchemy==2.0.30
pymysql==1.1.1
holidays==0.46
requests
